+++ app/navigation/BodyTopBarNavigation.js
@@ -3,30 +3,17 @@ import { Text, View, Button } from "react-native";
 import { NavigationContainer } from "@react-navigation/native";
 import { createMaterialTopTabNavigator } from "@react-navigation/material-top-tabs";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
+import { NutritionPage } from "../pages/NutritionPage";
+import { ComplicationsPage } from "../pages/ComplicationsPage";
+import { TestsPage } from "../pages/TestsPage";
 
 const TopTab = createMaterialTopTabNavigator();
-
-function MyPregnancy() {
-  return (
-    <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
-      <Text>My pregnancy tab</Text>
-    </View>
-  );
-}
-
-function Appointments() {
-  return (
-    <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
-      <Text>Appointments tab</Text>
-    </View>
-  );
-}
 
 function MyTopTabs() {
   //const insets = useSafeAreaInsets();
   return (
     <TopTab.Navigator
-      initialRouteName="MyPregnancy"
+      initialRouteName="Nutrition"
       tabBarOptions={{
         activeTintColor: "#91298D",
         labelStyle: { fontSize: 12 },
@@ -36,20 +23,26 @@ function MyTopTabs() {
       }}
     >
       <TopTab.Screen
-        name="MyPregnancy"
-        component={MyPregnancy}
-        options={{ tabBarLabel: "My Pregnancy" }}
+        name="Nutrition"
+        component={NutritionPage}
+        options={{ tabBarLabel: "Nutrition" }}
       />
 
       <TopTab.Screen
-        name="Appointments"
-        component={Appointments}
-        options={{ tabBarLabel: "Appointments" }}
+        name="Complications"
+        component={ComplicationsPage}
+        options={{ tabBarLabel: "Complications" }}
+      />
+
+      <TopTab.Screen
+        name="Tests"
+        component={TestsPage}
+        options={{ tabBarLabel: "Tests" }}
       />
     </TopTab.Navigator>
   );
 }
 
-export default function HomeTopBarNavigator() {
+export default function BodyTopBarNavigator() {
   return <MyTopTabs />;
 }
