+++ app/pages/AppointmentsPage.js
@@ -1,0 +1,10 @@
+import React from "react";
+import { Text, View } from "react-native";
+
+export function AppointmentsPage() {
+  return (
+    <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
+      <Text>Appointments tab</Text>
+    </View>
+  );
+}
