+++ app/pages/FacilitiesPage.js
@@ -1,0 +1,10 @@
+import React from "react";
+import { Text, View } from "react-native";
+
+export function FacilitiesPage() {
+  return (
+    <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
+      <Text>Facilities tab</Text>
+    </View>
+  );
+}
