+++ app/pages/MyPregnancyPage.js
@@ -1,0 +1,10 @@
+import React from "react";
+import { Text, View } from "react-native";
+
+export function MyPregnancyPage() {
+  return (
+    <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
+      <Text>My pregnancy tab</Text>
+    </View>
+  );
+}
